Доработала сообщение ошибки для формы для гостей и создала отдельные модули для прайса и комнат
