+++ css/style.css
@@ -1048,3 +1048,7 @@ fieldset {
 .hidden {
   display: none;
 }
+
+.pristine-error {
+  display: block;
+}
